Remove models & time parsing from `_client.py`
`_client.py` has gotten pretty big and hard to navigate, so this moves several bits around and does some cleanup:

- Encapsulate code for parsing/serializing timestamps and URLs into functions.
- Move above code into `_utils` module.
- Move `CdxRecord` and `Memento` into separate `_models` module.

Ultimately, this is all under the hood, and should not affect the public API.
