+++ wayback/__init__.py
@@ -2,10 +2,13 @@ from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-from ._client import (  # noqa
+from ._utils import memento_url_data  # noqa
+
+from ._models import (  # noqa
     CdxRecord,
-    Memento,
-    memento_url_data,
+    Memento)
+
+from ._client import (  # noqa
     Mode,
     WaybackClient,
     WaybackSession)
